Move Color to own file
Move druid Color impls over
