+++ src/guideline.rs
@@ -2,8 +2,9 @@ use serde::de::{Deserialize, Deserializer};
 use serde::ser::{Serialize, SerializeStruct, Serializer};
 use serde::{de, ser};
 
+use crate::color::Color;
 use crate::identifier::Identifier;
-use crate::shared_types::{Color, Plist};
+use crate::shared_types::Plist;
 
 /// A guideline associated with a glyph.
 #[derive(Debug, Clone, PartialEq)]
